Incluyo las funciones de tests. Uso long long.

+++ tp1/ej2/ej2-nuevo/balanza.h
@@ -5,7 +5,7 @@
 class Balanza{
 	public:
 		//inicializacion es O(sqrt(p)) si no es todo una mentira
-		Balanza(int p); 
+		Balanza(long long p); 
 		~Balanza();
 		void imprimir(FILE*); 
 		//puedo imprimir por consola o a un archivo
